Add files via upload
AtTiny85 PWM fan control with thermister and a LED on pin#2 for over temp alert. The fan starts at full power for 3 seconds and then idles down to 20%. can be adjusted in code. used a 50n06 mofet for driving fan at 12v and ran power to fan and only pwm the ground.